Moved Pydantic models to their own file.
