+++ src/app/main.py
@@ -1,7 +1,6 @@
 # Import the FastAPI module
 from fastapi import FastAPI
-from pydantic import BaseModel, Field
-from typing import Any
+from models.models import Match, Move
 
 # Instantiate the web application
 app = FastAPI()
@@ -9,16 +8,6 @@ app = FastAPI()
 # Endpoints operations
 
 # 1) /move endoint
-class Square(BaseModel):
-    x : int = Field(ge=0, le=2)
-    y : int = Field(ge=0, le=2)
-
-class Move(BaseModel):
-    matchId : int
-    playerId : str
-    # Literal["created_at", "updated_at"] = "created_at"
-    square: Square
-
 @app.post("/move", status_code=200)
 async def move(
     move : Move
@@ -26,9 +15,6 @@ async def move(
     return move
 
 # 2) /status
-class Match(BaseModel):
-    id : int
-
 @app.get("/status", status_code=200, response_model=Match)
 async def status(
     matchId : int
